undo some misguided changes + massive metric type cleanup and docs polish
